Reorganize tests to match folder structure of tested components
